+++ bin/codepatrol-pi-executor.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { runPiExecutorCli } from "../dist/src/executors/pi.js";
+
+process.exitCode = await runPiExecutorCli();